Client works

+++ lib/cli/daemon.go
@@ -5,6 +5,7 @@ import (
 	"net/http"
 	"os"
 	"path/filepath"
+	"sort"
 
 	"github.com/uber/makisu/lib/log"
 	"github.com/uber/makisu/lib/mountutils"
@@ -12,24 +13,24 @@ import (
 )
 
 type DaemonApplication struct {
-	Socket  string `commander:"flag=socket,The path to the socket that Kodder will listen on for build requests"`
-	Replace bool   `commander:"flag=replace,Whether or not to remove an existing socket at the same path"`
-	Port    *int   `commander:"flag=port,The port that Kodder will listen on for build requests"`
+	Replace bool    `commander:"flag=replace,Whether or not to remove an existing socket at the same path"`
+	Port    int     `commander:"flag=port,The port that Kodder will listen on for build requests"`
+	Socket  *string `commander:"flag=socket,The path to the socket that Kodder will listen on for build requests"`
 
 	building *atomic.Bool
 
-	// originals is the list of files/directories that were
+	// originals is the list of files that were
 	// there when the app was started.
 	originals map[string]bool
 }
 
 func NewDaemonApplication() *DaemonApplication {
 	return &DaemonApplication{
-		Socket:  defaultEnv("KODDER_SOCKET", "/kodder/kodder.sock"),
+		Port:    3456,
 		Replace: false,
 
 		building:  atomic.NewBool(false),
-		originals: map[string]bool{"/makisu-storage": true},
+		originals: map[string]bool{},
 	}
 }
 
@@ -61,6 +62,7 @@ func (app *DaemonApplication) listen() error {
 }
 
 func (app *DaemonApplication) cleanup() error {
+	log.Infof("Cleaning up filesystem")
 	root, err := os.Open("/")
 	if err != nil {
 		return fmt.Errorf("cleanup error: %v", err)
@@ -69,21 +71,68 @@ func (app *DaemonApplication) cleanup() error {
 	if err != nil {
 		return fmt.Errorf("cleanup error: %v", err)
 	}
+
+	trash := map[string]bool{}
 	for _, info := range infos {
 		fname := filepath.Join("/", info.Name())
-		if skip, err := mountutils.ContainsMountpoint(fname); err != nil {
+		if _, found := RootLevelSkips[fname]; found {
+			continue
+		} else if skip, err := mountutils.IsMounted(fname); err != nil {
 			return fmt.Errorf("failed to check mountpoints: %v", err)
 		} else if skip {
-			log.Debugf("Skipping cleanup of %v, mountpoint detected", fname)
-			continue
-		} else if _, found := app.originals[fname]; found {
-			log.Debugf("Skipping cleanup of %v, present at container launch", fname)
+			log.Infof("Skipping cleanup of %v, mountpoint detected", fname)
 			continue
 		}
 
-		log.Infof("Cleaning up dir: %v", fname)
-		if err := os.RemoveAll(fname); err != nil {
-			return fmt.Errorf("failed to cleanup %v: %v", fname, err)
+		fi, err := os.Stat(fname)
+		if err != nil {
+			return fmt.Errorf("failed to stat file to remove %v: %v", fname, err)
+		} else if _, found := app.originals[fname]; found && !fi.IsDir() {
+			log.Infof("Skipping cleanup of %v, present at container launch", fname)
+			continue
+		}
+
+		if !fi.IsDir() {
+			if err := os.RemoveAll(fname); err != nil {
+				return fmt.Errorf("failed to cleanup %v: %v", fname, err)
+			}
+			continue
+		}
+
+		err = filepath.Walk(fname, func(path string, fi os.FileInfo, err error) error {
+			if _, found := app.originals[path]; found {
+				return nil
+			} else if mounted, _ := mountutils.IsMounted(path); mounted {
+				log.Infof("Path mounted: %v", path)
+				return nil
+			} else if mountpoint, _ := mountutils.IsMountpoint(path); mountpoint {
+				log.Infof("Mountpoint detected: %v", path)
+				return filepath.SkipDir
+			} else if contains, _ := mountutils.ContainsMountpoint(path); contains {
+				log.Infof("Inner mountpoint detected: %v", path)
+				return nil
+			}
+			trash[path] = true
+			return nil
+		})
+		if err != nil {
+			return fmt.Errorf("failed to cleanup directory: %v", err)
+		}
+	}
+
+	for _, fname := range MustRemove {
+		trash[fname] = true
+	}
+
+	trashSlice := []string{}
+	for fname := range trash {
+		trashSlice = append(trashSlice, fname)
+	}
+	sort.Strings(trashSlice)
+	for i := len(trashSlice) - 1; i >= 0; i-- {
+		fname := trashSlice[i]
+		if err := os.RemoveAll(fname); err != nil && !os.IsNotExist(err) {
+			return fmt.Errorf("failed to cleanup directory: %v", err)
 		}
 	}
 	return nil
